Adding Athletes working, added background
Add Athletes function working, launches new activity using intents and
passes data back. Also added experimental background just because I was
getting bored of plain white, but it can be changed very easily in the
future. Potentially add a gradient effect to the listview so the lower
part of the text isn't obscured by the background? More to come
tomorrow.