Added a dialog for deletion
Also other slight changes. Eric, if you're reading this, I've decided to
continue working on this so that we might one day be able to put it on
the Play Store. While it's not great right now, I can definitely see it
as a potentially useful app once we put more time into it, and even a
potential money maker. I'll probably just be working on it over breaks
and stuff. Let me know if you want to help so we can talk
design/planning.